Struct google_gan1_beta1::Advertiser

pub struct Advertiser {
    pub status: Option<String>,
    pub product_feeds_enabled: Option<bool>,
    pub contact_phone: Option<String>,
    pub description: Option<String>,
    pub merchant_center_ids: Option<Vec<String>>,
    pub commission_duration: Option<i32>,
    pub contact_email: Option<String>,
    pub join_date: Option<String>,
    pub id: Option<String>,
    pub category: Option<String>,
    pub default_link_id: Option<String>,
    pub kind: Option<String>,
    pub site_url: Option<String>,
    pub name: Option<String>,
    pub payout_rank: Option<String>,
    pub epc_seven_day_average: Option<Money>,
    pub epc_ninety_day_average: Option<Money>,
    pub allow_publisher_created_links: Option<bool>,
    pub item: Option<Option<Box<Advertiser>>>,
    pub logo_url: Option<String>,
    pub redirect_domains: Option<Vec<String>>,
}

An AdvertiserResource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list advertisers (none)

get advertisers (response)

Fields

status

The status of the requesting publisher's relationship this advertiser.

product_feeds_enabled

Allows advertisers to submit product listings to Google Product Search.

contact_phone

Phone that this advertiser would like publishers to contact them with.

description

Description of the website the advertiser advertises from.

merchant_center_ids

List of merchant center ids for this advertiser

commission_duration

The longest possible length of a commission (how long the cookies on the customer's browser last before they expire).

contact_email

Email that this advertiser would like publishers to contact them with.

join_date

Date that this advertiser was approved as a Google Affiliate Network advertiser.

id

The ID of this advertiser.

category

Category that this advertiser belongs to. A valid list of categories can be found here: http://www.google.com/support/affiliatenetwork/advertiser/bin/answer.py?hl=en&answer=107581

default_link_id

The default link id for this advertiser.

kind

The kind for an advertiser.

site_url

URL of the website this advertiser advertises from.

name

The name of this advertiser.

payout_rank

A rank based on commissions paid to publishers over the past 90 days. A number between 1 and 4 where 4 means the top quartile (most money paid) and 1 means the bottom quartile (least money paid).

epc_seven_day_average

The sum of fees paid to publishers divided by the total number of clicks over the past seven days. This value should be multiplied by 100 at the time of display.

epc_ninety_day_average

The sum of fees paid to publishers divided by the total number of clicks over the past three months. This value should be multiplied by 100 at the time of display.

allow_publisher_created_links

True if the advertiser allows publisher created links, otherwise false.

item

The requested advertiser.

logo_url

URL to the logo this advertiser uses on the Google Affiliate Network.

redirect_domains

List of redirect URLs for this advertiser