Struct google_gan1_beta1::AdvertiserListCall

pub struct AdvertiserListCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Retrieves data about all advertisers that the requesting advertiser/publisher has access to.

A builder for the list method supported by a advertiser resource. It is not used directly, but through a AdvertiserMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.advertisers().list("role", "roleId")
             .relationship_status("Lorem")
             .page_token("et")
             .min_seven_day_epc(0.877582930428)
             .min_payout_rank(-45)
             .min_ninety_day_epc(0.427055674413)
             .max_results(61)
             .advertiser_category("kasd")
             .doit();

Methods

impl<'a, C, A> AdvertiserListCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, Advertisers)>

Perform the operation you have build so far.

fn role(self, new_value: &str) -> AdvertiserListCall<'a, C, A>

The role of the requester. Valid values: 'advertisers' or 'publishers'.

Sets the role path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn role_id(self, new_value: &str) -> AdvertiserListCall<'a, C, A>

The ID of the requesting advertiser or publisher.

Sets the role id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn relationship_status(self, new_value: &str) -> AdvertiserListCall<'a, C, A>

Filters out all advertisers for which do not have the given relationship status with the requesting publisher.

Sets the relationship status query property to the given value.

fn page_token(self, new_value: &str) -> AdvertiserListCall<'a, C, A>

The value of 'nextPageToken' from the previous page. Optional.

Sets the page token query property to the given value.

fn min_seven_day_epc(self, new_value: f64) -> AdvertiserListCall<'a, C, A>

Filters out all advertisers that have a seven day EPC average lower than the given value (inclusive). Min value: 0.0. Optional.

Sets the min seven day epc query property to the given value.

fn min_payout_rank(self, new_value: i32) -> AdvertiserListCall<'a, C, A>

A value between 1 and 4, where 1 represents the quartile of advertisers with the lowest ranks and 4 represents the quartile of advertisers with the highest ranks. Filters out all advertisers with a lower rank than the given quartile. For example if a 2 was given only advertisers with a payout rank of 25 or higher would be included. Optional.

Sets the min payout rank query property to the given value.

fn min_ninety_day_epc(self, new_value: f64) -> AdvertiserListCall<'a, C, A>

Filters out all advertisers that have a ninety day EPC average lower than the given value (inclusive). Min value: 0.0. Optional.

Sets the min ninety day epc query property to the given value.

fn max_results(self, new_value: u32) -> AdvertiserListCall<'a, C, A>

Max number of items to return in this page. Optional. Defaults to 20.

Sets the max results query property to the given value.

fn advertiser_category(self, new_value: &str) -> AdvertiserListCall<'a, C, A>

Caret() delimted list of advertiser categories. Valid categories are defined here: http://www.google.com/support/affiliatenetwork/advertiser/bin/answer.py?hl=en&answer=107581. Filters out all advertisers not in one of the given advertiser categories. Optional.

Sets the advertiser category query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> AdvertiserListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> AdvertiserListCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.