Struct google_gan1_beta1::CcOfferRewards

pub struct CcOfferRewards {
    pub category: Option<String>,
    pub amount: Option<f64>,
    pub min_reward_tier: Option<f64>,
    pub additional_details: Option<String>,
    pub max_reward_tier: Option<f64>,
    pub expiration_months: Option<f64>,
}

For cards with rewards programs, detailed rules about how the program works.

This type is not used in any activity, and only used as part of another schema.

Fields
category: Option<String>

The kind of purchases covered by this rule.

amount: Option<f64>

The number of units rewarded per purchase dollar.

min_reward_tier: Option<f64>

The minimum purchase amount in the given category before this rule applies.

additional_details: Option<String>

Other limits, for example, if this rule only applies during an introductory period.

max_reward_tier: Option<f64>

The maximum purchase amount in the given category for this rule to apply.

expiration_months: Option<f64>

How long rewards granted by this rule last.