Struct google_gan1_beta1::Event

pub struct Event {
    pub order_id: Option<String>,
    pub status: Option<String>,
    pub modify_date: Option<String>,
    pub advertiser_name: Option<String>,
    pub charge_id: Option<String>,
    pub charge_type: Option<String>,
    pub publisher_name: Option<String>,
    pub event_date: Option<String>,
    pub network_fee: Option<Money>,
    pub advertiser_id: Option<String>,
    pub kind: Option<String>,
    pub publisher_id: Option<String>,
    pub member_id: Option<String>,
    pub publisher_fee: Option<Money>,
    pub earnings: Option<Money>,
    pub products: Option<Vec<EventProducts>>,
    pub commissionable_sales: Option<Money>,
    pub type_: Option<String>,
}

An EventResource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list events (none)

Fields

order_id

The order ID for this event. Only returned for conversion events.

status

Status of the event (active|canceled). Only returned for charge and conversion events.

modify_date

The date-time this event was last modified as a RFC 3339 date-time value.

advertiser_name

The name of the advertiser for this event.

charge_id

The charge ID for this event. Only returned for charge events.

charge_type

Charge type of the event (other|slotting_fee|monthly_minimum|tier_bonus|debit|credit). Only returned for charge events.

publisher_name

The name of the publisher for this event.

event_date

The date-time this event was initiated as a RFC 3339 date-time value.

network_fee

Fee that the advertiser paid to the Google Affiliate Network.

advertiser_id

The ID of advertiser for this event.

kind

The kind for one event.

publisher_id

The ID of the publisher for this event.

member_id

The ID of the member attached to this event. Only returned for conversion events.

publisher_fee

Fee that the advertiser paid to the publisher.

earnings

Earnings by the publisher.

products

Products associated with the event.

commissionable_sales

Amount of money exchanged during the transaction. Only returned for charge and conversion events.

type_

Type of the event (action|transaction|charge).