Struct google_gan1_beta1::EventListCall

pub struct EventListCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Retrieves event data for a given advertiser/publisher.

A builder for the list method supported by a event resource. It is not used directly, but through a EventMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.events().list("role", "roleId")
             .type_("At")
             .status("labore")
             .sku("invidunt")
             .publisher_id("ea")
             .product_category("sadipscing")
             .page_token("rebum.")
             .order_id("dolore")
             .modify_date_min("nonumy")
             .modify_date_max("sed")
             .member_id("aliquyam")
             .max_results(48)
             .link_id("eirmod")
             .event_date_min("consetetur")
             .event_date_max("labore")
             .charge_type("sed")
             .advertiser_id("ea")
             .doit();

Methods

impl<'a, C, A> EventListCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, Events)>

Perform the operation you have build so far.

fn role(self, new_value: &str) -> EventListCall<'a, C, A>

The role of the requester. Valid values: 'advertisers' or 'publishers'.

Sets the role path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn role_id(self, new_value: &str) -> EventListCall<'a, C, A>

The ID of the requesting advertiser or publisher.

Sets the role id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn type_(self, new_value: &str) -> EventListCall<'a, C, A>

Filters out all events that are not of the given type. Valid values: 'action', 'transaction', 'charge'. Optional.

Sets the type query property to the given value.

fn status(self, new_value: &str) -> EventListCall<'a, C, A>

Filters out all events that do not have the given status. Valid values: 'active', 'canceled'. Optional.

Sets the status query property to the given value.

fn sku(self, new_value: &str) -> EventListCall<'a, C, A>

Caret() delimited list of SKUs. Filters out all events that do not reference one of the given SKU. Optional.

Sets the sku query property to the given value.

fn publisher_id(self, new_value: &str) -> EventListCall<'a, C, A>

Caret() delimited list of publisher IDs. Filters out all events that do not reference one of the given publishers IDs. Only used when under advertiser role. Optional.

Sets the publisher id query property to the given value.

fn product_category(self, new_value: &str) -> EventListCall<'a, C, A>

Caret() delimited list of product categories. Filters out all events that do not reference a product in one of the given product categories. Optional.

Sets the product category query property to the given value.

fn page_token(self, new_value: &str) -> EventListCall<'a, C, A>

The value of 'nextPageToken' from the previous page. Optional.

Sets the page token query property to the given value.

fn order_id(self, new_value: &str) -> EventListCall<'a, C, A>

Caret() delimited list of order IDs. Filters out all events that do not reference one of the given order IDs. Optional.

Sets the order id query property to the given value.

fn modify_date_min(self, new_value: &str) -> EventListCall<'a, C, A>

Filters out all events modified earlier than given date. Optional. Defaults to 24 hours before the current modifyDateMax, if modifyDateMax is explicitly set.

Sets the modify date min query property to the given value.

fn modify_date_max(self, new_value: &str) -> EventListCall<'a, C, A>

Filters out all events modified later than given date. Optional. Defaults to 24 hours after modifyDateMin, if modifyDateMin is explicitly set.

Sets the modify date max query property to the given value.

fn member_id(self, new_value: &str) -> EventListCall<'a, C, A>

Caret() delimited list of member IDs. Filters out all events that do not reference one of the given member IDs. Optional.

Sets the member id query property to the given value.

fn max_results(self, new_value: u32) -> EventListCall<'a, C, A>

Max number of offers to return in this page. Optional. Defaults to 20.

Sets the max results query property to the given value.

fn link_id(self, new_value: &str) -> EventListCall<'a, C, A>

Caret() delimited list of link IDs. Filters out all events that do not reference one of the given link IDs. Optional.

Sets the link id query property to the given value.

fn event_date_min(self, new_value: &str) -> EventListCall<'a, C, A>

Filters out all events earlier than given date. Optional. Defaults to 24 hours from current date/time.

Sets the event date min query property to the given value.

fn event_date_max(self, new_value: &str) -> EventListCall<'a, C, A>

Filters out all events later than given date. Optional. Defaults to 24 hours after eventMin.

Sets the event date max query property to the given value.

fn charge_type(self, new_value: &str) -> EventListCall<'a, C, A>

Filters out all charge events that are not of the given charge type. Valid values: 'other', 'slotting_fee', 'monthly_minimum', 'tier_bonus', 'credit', 'debit'. Optional.

Sets the charge type query property to the given value.

fn advertiser_id(self, new_value: &str) -> EventListCall<'a, C, A>

Caret() delimited list of advertiser IDs. Filters out all events that do not reference one of the given advertiser IDs. Only used when under publishers role. Optional.

Sets the advertiser id query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> EventListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> EventListCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.