Struct google_gan1_beta1::LinkMethods

pub struct LinkMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on link resources. It is not used directly, but through the Gan hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_gan1_beta1 as gan1_beta1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use gan1_beta1::Gan;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Gan::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `insert(...)` and `list(...)`
// to build up your call.
let rb = hub.links();

Methods

impl<'a, C, A> LinkMethods<'a, C, A>

fn insert(&self, request: Link, role: &str, role_id: &str) -> LinkInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Inserts a new link.

Arguments

request - No description provided.

role - The role of the requester. Valid values: 'advertisers' or 'publishers'.

roleId - The ID of the requesting advertiser or publisher.

fn get(&self, role: &str, role_id: &str, link_id: &str) -> LinkGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves data about a single link if the requesting advertiser/publisher has access to it. Advertisers can look up their own links. Publishers can look up visible links or links belonging to advertisers they are in a relationship with.

Arguments

role - The role of the requester. Valid values: 'advertisers' or 'publishers'.

roleId - The ID of the requesting advertiser or publisher.

linkId - The ID of the link to look up.

fn list(&self, role: &str, role_id: &str) -> LinkListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves all links that match the query parameters.

Arguments

role - The role of the requester. Valid values: 'advertisers' or 'publishers'.

roleId - The ID of the requesting advertiser or publisher.