Struct google-gan1_beta1::Publisher

pub struct Publisher {
    pub status: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub classification: Option<String>,
    pub epc_seven_day_average: Option<Money>,
    pub sites: Vec<String>,
    pub epc_ninety_day_average: Option<Money>,
    pub item: Option<Box<Publisher>>,
    pub payout_rank: Option<String>,
    pub join_date: Option<String>,
    pub id: Option<String>,
}

A PublisherResource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list publishers (none)

get publishers (response)

Fields

status

The status of the requesting advertiser's relationship with this publisher.

kind

The kind for a publisher.

name

The name of this publisher.

classification

Classification that this publisher belongs to. See this link for all publisher classifications: http://www.google.com/support/affiliatenetwork/advertiser/bin/answer.py?hl=en&answer=107625&ctx=cb&src=cb&cbid=-k5fihzthfaik&cbrank=4

epc_seven_day_average

The sum of fees paid to this publisher divided by the total number of clicks over the past seven days. Values are multiplied by 100 for display purposes.

sites

Websites that this publisher uses to advertise.

epc_ninety_day_average

The sum of fees paid to this publisher divided by the total number of clicks over the past three months. Values are multiplied by 100 for display purposes.

item

The requested publisher.

payout_rank

A rank based on commissions paid to this publisher over the past 90 days. A number between 1 and 4 where 4 means the top quartile (most money paid) and 1 means the bottom quartile (least money paid).

join_date

Date that this publisher was approved as a Google Affiliate Network publisher.

id

The ID of this publisher.