Struct google_gan1_beta1::Report

pub struct Report {
    pub totals_rows: Option<Vec<Vec<String>>>,
    pub kind: Option<String>,
    pub rows: Option<Vec<Vec<String>>>,
    pub end_date: Option<String>,
    pub matching_row_count: Option<String>,
    pub column_names: Option<Vec<String>>,
    pub type_: Option<String>,
    pub start_date: Option<String>,
}

A ReportResource representing a report of a certain type either for an advertiser or publisher.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get reports (response)

Fields
totals_rows: Option<Vec<Vec<String>>>

The totals rows for the report

kind: Option<String>

The kind for a report.

rows: Option<Vec<Vec<String>>>

The rows of data for the report

end_date: Option<String>

The end of the date range for this report, exclusive.

matching_row_count: Option<String>

The number of matching rows before paging is applied.

column_names: Option<Vec<String>>

The column names for the report

type_: Option<String>

The report type.

start_date: Option<String>

The start of the date range for this report, inclusive.