Struct google_gan1_beta1::CcOffer

pub struct CcOffer {
    pub luggage_insurance: Option<String>,
    pub credit_limit_min: Option<f64>,
    pub card_name: Option<String>,
    pub credit_limit_max: Option<f64>,
    pub grace_period_display: Option<String>,
    pub additional_card_holder_fee: Option<String>,
    pub offer_id: Option<String>,
    pub min_purchase_rate: Option<f64>,
    pub card_benefits: Option<Vec<String>>,
    pub default_fees: Option<Vec<CcOfferDefaultFees>>,
    pub offers_immediate_cash_reward: Option<bool>,
    pub travel_insurance: Option<String>,
    pub returned_payment_fee: Option<String>,
    pub issuer: Option<String>,
    pub existing_customer_only: Option<bool>,
    pub annual_fee_display: Option<String>,
    pub initial_setup_and_processing_fee: Option<String>,
    pub minimum_finance_charge: Option<String>,
    pub purchase_rate_additional_details: Option<String>,
    pub reward_unit: Option<String>,
    pub credit_rating_display: Option<String>,
    pub fraud_liability: Option<String>,
    pub cash_advance_terms: Option<String>,
    pub landing_page_url: Option<String>,
    pub intro_cash_advance_terms: Option<String>,
    pub first_year_annual_fee: Option<f64>,
    pub intro_purchase_terms: Option<String>,
    pub rewards: Option<Vec<CcOfferRewards>>,
    pub extended_warranty: Option<String>,
    pub max_purchase_rate: Option<f64>,
    pub tracking_url: Option<String>,
    pub late_payment_fee: Option<String>,
    pub over_limit_fee: Option<String>,
    pub card_type: Option<String>,
    pub approved_categories: Option<Vec<String>>,
    pub reward_partner: Option<String>,
    pub intro_balance_transfer_terms: Option<String>,
    pub annual_reward_maximum: Option<f64>,
    pub foreign_currency_transaction_fee: Option<String>,
    pub annual_fee: Option<f64>,
    pub car_rental_insurance: Option<String>,
    pub variable_rates_update_frequency: Option<String>,
    pub kind: Option<String>,
    pub additional_card_benefits: Option<Vec<String>>,
    pub issuer_website: Option<String>,
    pub balance_computation_method: Option<String>,
    pub apr_display: Option<String>,
    pub image_url: Option<String>,
    pub prohibited_categories: Option<Vec<String>>,
    pub variable_rates_last_updated: Option<String>,
    pub age_minimum: Option<f64>,
    pub network: Option<String>,
    pub purchase_rate_type: Option<String>,
    pub statement_copy_fee: Option<String>,
    pub rewards_have_blackout_dates: Option<bool>,
    pub emergency_insurance: Option<String>,
    pub flight_accident_insurance: Option<String>,
    pub rewards_expire: Option<bool>,
    pub balance_transfer_terms: Option<String>,
    pub disclaimer: Option<String>,
    pub issuer_id: Option<String>,
    pub age_minimum_details: Option<String>,
    pub bonus_rewards: Option<Vec<CcOfferBonusRewards>>,
}

A credit card offer. There are many possible result fields. We provide two different views of the data, or "projections." The "full" projection includes every result field. And the "summary" projection, which is the default, includes a smaller subset of the fields. The fields included in the summary projection are marked as such in their descriptions.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list cc offers (none)

Fields
luggage_insurance: Option<String>

If you get coverage when you use the card for the given activity, this field describes it.

credit_limit_min: Option<f64>

The low end for credit limits the issuer imposes on recipients of this card.

card_name: Option<String>

The issuer's name for the card, including any trademark or service mark designators. A summary field.

credit_limit_max: Option<f64>

The high end for credit limits the issuer imposes on recipients of this card.

grace_period_display: Option<String>

Text describing the grace period before finance charges apply. A summary field.

additional_card_holder_fee: Option<String>

Any extra fees levied on card holders.

offer_id: Option<String>

This offer's ID. A summary field.

min_purchase_rate: Option<f64>

The lowest interest rate the issuer charges on this card. Expressed as an absolute number, not as a percentage.

card_benefits: Option<Vec<String>>

A list of what the issuer thinks are the most important benefits of the card. Usually summarizes the rewards program, if there is one. A summary field.

default_fees: Option<Vec<CcOfferDefaultFees>>

Fees for defaulting on your payments.

offers_immediate_cash_reward: Option<bool>

Whether a cash reward program lets you get cash back sooner than end of year or other longish period.

travel_insurance: Option<String>

If you get coverage when you use the card for the given activity, this field describes it.

returned_payment_fee: Option<String>

Text describing the fee for a payment that doesn't clear. A summary field.

issuer: Option<String>

Name of card issuer. A summary field.

existing_customer_only: Option<bool>

Whether this card is only available to existing customers of the issuer.

annual_fee_display: Option<String>

Text describing the annual fee, including any difference for the first year. A summary field.

initial_setup_and_processing_fee: Option<String>

Fee for setting up the card.

minimum_finance_charge: Option<String>

Text describing how much missing the grace period will cost.

purchase_rate_additional_details: Option<String>

Text describing any additional details for the purchase rate. A summary field.

reward_unit: Option<String>

For cards with rewards programs, the unit of reward. For example, miles, cash back, points.

credit_rating_display: Option<String>

Text describing the credit ratings required for recipients of this card, for example "Excellent/Good." A summary field.

fraud_liability: Option<String>

If you get coverage when you use the card for the given activity, this field describes it.

cash_advance_terms: Option<String>

Text describing the terms for cash advances. A summary field.

landing_page_url: Option<String>

The link to the issuer's page for this card. A summary field.

intro_cash_advance_terms: Option<String>

Text describing the terms for introductory period cash advances. A summary field.

first_year_annual_fee: Option<f64>

The annual fee for the first year, if different from the ongoing fee. Optional.

intro_purchase_terms: Option<String>

Text describing the terms for introductory period purchases. A summary field.

rewards: Option<Vec<CcOfferRewards>>

For cards with rewards programs, detailed rules about how the program works.

extended_warranty: Option<String>

If you get coverage when you use the card for the given activity, this field describes it.

max_purchase_rate: Option<f64>

The highest interest rate the issuer charges on this card. Expressed as an absolute number, not as a percentage.

tracking_url: Option<String>

The link to ping to register a click on this offer. A summary field.

late_payment_fee: Option<String>

Text describing how much a late payment will cost, eg "up to $35." A summary field.

over_limit_fee: Option<String>

Fee for exceeding the card's charge limit.

card_type: Option<String>

What kind of credit card this is, for example secured or unsecured.

approved_categories: Option<Vec<String>>

Possible categories for this card, eg "Low Interest" or "Good." A summary field.

reward_partner: Option<String>

The company that redeems the rewards, if different from the issuer.

intro_balance_transfer_terms: Option<String>

Text describing the terms for introductory period balance transfers. A summary field.

annual_reward_maximum: Option<f64>

The largest number of units you may accumulate in a year.

foreign_currency_transaction_fee: Option<String>

Fee for each transaction involving a foreign currency.

annual_fee: Option<f64>

The ongoing annual fee, in dollars.

car_rental_insurance: Option<String>

If you get coverage when you use the card for the given activity, this field describes it.

variable_rates_update_frequency: Option<String>

How often variable rates are updated.

kind: Option<String>

The kind for one credit card offer. A summary field.

additional_card_benefits: Option<Vec<String>>

More marketing copy about the card's benefits. A summary field.

issuer_website: Option<String>

The generic link to the issuer's site.

balance_computation_method: Option<String>

Text describing how the balance is computed. A summary field.

apr_display: Option<String>

Text describing the purchase APR. A summary field.

image_url: Option<String>

The link to the image of the card that is shown on Connect Commerce. A summary field.

prohibited_categories: Option<Vec<String>>

Categories in which the issuer does not wish the card to be displayed. A summary field.

variable_rates_last_updated: Option<String>

When variable rates were last updated.

age_minimum: Option<f64>

The youngest a recipient of this card may be.

network: Option<String>

Which network (eg Visa) the card belongs to. A summary field.

purchase_rate_type: Option<String>

Fixed or variable.

statement_copy_fee: Option<String>

Fee for requesting a copy of your statement.

rewards_have_blackout_dates: Option<bool>

For airline miles rewards, tells whether blackout dates apply to the miles.

emergency_insurance: Option<String>

If you get coverage when you use the card for the given activity, this field describes it.

flight_accident_insurance: Option<String>

If you get coverage when you use the card for the given activity, this field describes it.

rewards_expire: Option<bool>

Whether accumulated rewards ever expire.

balance_transfer_terms: Option<String>

Text describing the terms for balance transfers. A summary field.

disclaimer: Option<String>

A notice that, if present, is referenced via an asterisk by many of the other summary fields. If this field is present, it will always start with an asterisk ("*"), and must be prominently displayed with the offer. A summary field.

issuer_id: Option<String>

The Google Affiliate Network ID of the advertiser making this offer.

age_minimum_details: Option<String>

Text describing the details of the age minimum restriction.

bonus_rewards: Option<Vec<CcOfferBonusRewards>>

For cards with rewards programs, extra circumstances whereby additional rewards may be granted.