Struct google_gan1_beta1::Link

pub struct Link {
    pub link_type: Option<String>,
    pub start_date: Option<String>,
    pub end_date: Option<String>,
    pub description: Option<String>,
    pub availability: Option<String>,
    pub image_alt_text: Option<String>,
    pub advertiser_id: Option<String>,
    pub click_tracking_url: Option<String>,
    pub promotion_type: Option<String>,
    pub duration: Option<String>,
    pub authorship: Option<String>,
    pub id: Option<String>,
    pub is_active: Option<bool>,
    pub destination_url: Option<String>,
    pub special_offers: Option<LinkSpecialOffers>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub epc_seven_day_average: Option<Money>,
    pub create_date: Option<String>,
    pub epc_ninety_day_average: Option<Money>,
    pub impression_tracking_url: Option<String>,
}

A LinkResource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert links (request|response)

get links (response)

list links (none)

Fields
link_type: Option<String>

The link type.

start_date: Option<String>

Date that this link becomes active.

end_date: Option<String>

Date that this link becomes inactive.

description: Option<String>

Description.

availability: Option<String>

Availability.

image_alt_text: Option<String>

image alt text.

advertiser_id: Option<String>

The advertiser id for the advertiser who owns this link.

click_tracking_url: Option<String>

Tracking url for clicks.

promotion_type: Option<String>

Promotion Type

duration: Option<String>

Duration

authorship: Option<String>

Authorship

id: Option<String>

The ID of this link.

is_active: Option<bool>

Flag for if this link is active.

destination_url: Option<String>

The destination URL for the link.

special_offers: Option<LinkSpecialOffers>

Special offers on the link.

kind: Option<String>

The kind for one entity.

name: Option<String>

The logical name for this link.

epc_seven_day_average: Option<Money>

The sum of fees paid to publishers divided by the total number of clicks over the past seven days on this link. This value should be multiplied by 100 at the time of display.

create_date: Option<String>

Date that this link was created.

epc_ninety_day_average: Option<Money>

The sum of fees paid to publishers divided by the total number of clicks over the past three months on this link. This value should be multiplied by 100 at the time of display.

impression_tracking_url: Option<String>

Tracking url for impressions.