Struct google_gan1_beta1::Link

pub struct Link {
    pub link_type: Option<String>,
    pub start_date: Option<String>,
    pub end_date: Option<String>,
    pub description: Option<String>,
    pub availability: Option<String>,
    pub image_alt_text: Option<String>,
    pub advertiser_id: Option<String>,
    pub click_tracking_url: Option<String>,
    pub promotion_type: Option<String>,
    pub duration: Option<String>,
    pub authorship: Option<String>,
    pub id: Option<String>,
    pub is_active: Option<bool>,
    pub destination_url: Option<String>,
    pub special_offers: Option<LinkSpecialOffers>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub epc_seven_day_average: Option<Money>,
    pub create_date: Option<String>,
    pub epc_ninety_day_average: Option<Money>,
    pub impression_tracking_url: Option<String>,
}

A LinkResource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert links (request|response)

get links (response)

list links (none)

Fields

link_type

The link type.

start_date

Date that this link becomes active.

end_date

Date that this link becomes inactive.

description

Description.

availability

Availability.

image_alt_text

image alt text.

advertiser_id

The advertiser id for the advertiser who owns this link.

click_tracking_url

Tracking url for clicks.

promotion_type

Promotion Type

duration

Duration

authorship

Authorship

id

The ID of this link.

is_active

Flag for if this link is active.

destination_url

The destination URL for the link.

special_offers

Special offers on the link.

kind

The kind for one entity.

name

The logical name for this link.

epc_seven_day_average

The sum of fees paid to publishers divided by the total number of clicks over the past seven days on this link. This value should be multiplied by 100 at the time of display.

create_date

Date that this link was created.

epc_ninety_day_average

The sum of fees paid to publishers divided by the total number of clicks over the past three months on this link. This value should be multiplied by 100 at the time of display.

impression_tracking_url

Tracking url for impressions.