Struct google_gan1_beta1::Report

pub struct Report {
    pub totals_rows: Option<Vec<Vec<String>>>,
    pub kind: Option<String>,
    pub rows: Option<Vec<Vec<String>>>,
    pub end_date: Option<String>,
    pub matching_row_count: Option<String>,
    pub column_names: Option<Vec<String>>,
    pub type_: Option<String>,
    pub start_date: Option<String>,
}

A ReportResource representing a report of a certain type either for an advertiser or publisher.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get reports (response)

Fields

totals_rows

The totals rows for the report

kind

The kind for a report.

rows

The rows of data for the report

end_date

The end of the date range for this report, exclusive.

matching_row_count

The number of matching rows before paging is applied.

column_names

The column names for the report

type_

The report type.

start_date

The start of the date range for this report, inclusive.